move the receiver package to within the service package
